Added a little cube for scale
The black hole is still a bit finicky.  I think this is because there are a lot of calculations being made, due to the lensing effect and the number of rings I added to make it look good.  My next commit aims to fix that by modifying the scripts used, like SGTRing and SGTLens, so hopefully the performance increases and the black holes stops flickering.