Added O2 Tank Collisions and Found Error
When the player interacts/collides with an O2 tank, it updates their Oxygen count.  It also deletes the object.  However, an error was found while testing this:
If the amount the player gets back from an O2 tank is still below the critical amount, another O2 tank will not spawn.  This means the player only gets one O2 tank, and no more.