Minor Audio Changes
I changed the background music to a newer song I made; I think it sounds a bit better than that really monotonous initial track I made.  It's got a long pause in it since I think it's better to scatter the music a bit instead of having a constant loop.  Also, I added a audio clip for when a fuel canister is picked up.  I haven't gotten that to work yet, but the audio is ready to be used now.